Move structs into main file
